+++ Libplanet.Action.Tests/Common/DetectRehearsal.cs
@@ -25,10 +25,10 @@ namespace Libplanet.Action.Tests.Common
         public override IWorld Execute(IActionContext context)
         {
             IWorld previousState = context.PreviousState;
-            IAccount legacyAccount = previousState.GetAccount(ReservedAddresses.LegacyAccount);
+            IAccount legacyAccount = previousState.GetAccount(ReservedAddresses.DefaultAccount);
             ResultState = context.Rehearsal;
             return previousState.SetAccount(
-                ReservedAddresses.LegacyAccount,
+                ReservedAddresses.DefaultAccount,
                 legacyAccount.SetState(
                     TargetAddress,
                     new Bencodex.Types.Boolean(context.Rehearsal)