+++ Libplanet.Explorer.Executable/Program.cs
@@ -147,6 +147,10 @@ consecutive blocks.")]
                 Description = "The name of MySQL database to use.")]
             string mysqlDatabase = null,
             [Option(
+                "fee-account-address",
+                Description = "The account address where fee stored.")]
+            string feeAccountAddress = null,
+            [Option(
                 "max-transactions-per-block",
                 Description = @"The number of maximum transactions able to be included
 in a block.")]
@@ -188,6 +192,7 @@ If omitted (default) explorer only the local blockchain store.")]
                 mysqlUsername,
                 mysqlPassword,
                 mysqlDatabase,
+                feeAccountAddress,
                 maxTransactionsPerBlock,
                 maxTransactionsBytes,
                 maxGenesisTransactionsBytes,
@@ -224,6 +229,7 @@ If omitted (default) explorer only the local blockchain store.")]
                         stateStore,
                         options.GetGenesisBlock(policy),
                         new ActionEvaluator(
+                            policy.SystemAccountsGetter,
                             _ => policy.BlockAction,
                             stateStore,
                             new SingleActionLoader(typeof(NullAction))));
@@ -396,6 +402,7 @@ If omitted (default) explorer only the local blockchain store.")]
         private static BlockPolicy LoadBlockPolicy(Options options)
         {
             return new BlockPolicy(
+                systemAccountsGetter: new SystemAccountsGetter(_ => options.FeeAccountAddress),
                 blockAction: null,
                 blockInterval: TimeSpan.FromMilliseconds(options.BlockIntervalMilliseconds),
                 getMaxTransactionsBytes: i => i > 0
@@ -438,6 +445,8 @@ If omitted (default) explorer only the local blockchain store.")]
             {
                 _impl = blockPolicy;
             }
+
+            public ISystemAccountsGetter SystemAccountsGetter => _impl.SystemAccountsGetter;
 
             public IAction BlockAction => _impl.BlockAction;