+++ Libplanet.Tests/Action/AccountDiffTest.cs
@@ -14,6 +14,9 @@ namespace Libplanet.Tests.Action
 {
     public class AccountDiffTest
     {
+        private readonly ISystemAccounts _systemAccounts = new SystemAccounts(
+            new SystemAccountsGetter(_ => ReservedAddresses.DefaultAccount), null);
+
         public AccountDiffTest()
         {
             USD = Currency.Uncapped("USD", 2, null);
@@ -146,6 +149,7 @@ namespace Libplanet.Tests.Action
                 signer,
                 null,
                 signer,
+                _systemAccounts,
                 0,
                 Block.CurrentProtocolVersion,
                 new World(