+++ Libplanet.Tests/Blockchain/Renderers/LoggedActionRendererTest.cs
@@ -23,6 +23,9 @@ namespace Libplanet.Tests.Blockchain.Renderers
     {
         private static IValue _action = new DumbAction().PlainValue;
 
+        private static ISystemAccounts _systemAccounts = new SystemAccounts(
+            new SystemAccountsGetter(_ => ReservedAddresses.DefaultAccount), null);
+
         private static IWorld _world = new World(new MockWorldState());
 
         private static Block _genesis =
@@ -73,8 +76,9 @@ namespace Libplanet.Tests.Blockchain.Renderers
                     default,
                     default,
                     default,
-                    Block.CurrentProtocolVersion,
+                    _systemAccounts,
                     123,
+                    Block.CurrentProtocolVersion,
                     _world,
                     default,
                     0,