+++ Libplanet.Tests/TestUtils.cs
@@ -607,9 +607,10 @@ Actual (C# array lit):   new byte[{actual.LongLength}] {{ {actualRepr} }}";
             };
 
             var actionEvaluator = new ActionEvaluator(
-                    _ => policy.BlockAction,
-                    stateStore: stateStore,
-                    actionTypeLoader: new SingleActionLoader(typeof(T)));
+                policy.SystemAccountsGetter,
+                _ => policy.BlockAction,
+                stateStore: stateStore,
+                actionTypeLoader: new SingleActionLoader(typeof(T)));
 
             if (genesisBlock is null)
             {